extract Roles helper class a la Sets/Maps
