+++ graylog2-server/src/main/java/org/graylog2/rest/resources/roles/RolesResource.java
@@ -35,6 +35,7 @@ import org.graylog2.security.InMemoryRolePermissionResolver;
 import org.graylog2.shared.rest.resources.RestResource;
 import org.graylog2.shared.security.RestPermissions;
 import org.graylog2.shared.users.Role;
+import org.graylog2.shared.users.Roles;
 import org.graylog2.users.RoleImpl;
 import org.graylog2.users.RoleService;
 import org.joda.time.DateTimeZone;
@@ -184,7 +185,7 @@ public class RolesResource extends RestResource {
             final Set<String> roleIds = user.getRoleIds();
             final Set<String> roleNames = Sets.newHashSet(
                     Collections2.transform(roleIds,
-                                           new Role.RoleIdToNameFunction(idMap)));
+                                           Roles.roleIdToNameFunction(idMap)));
             userSummaries.add(UserSummary.create(
                     user.getId(),
                     user.getName(),
